back-end/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
strawberry-graphql[fastapi]==0.220.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
